added file divide option

+++ DHGSystems.FileSortingWithLimitedMemory/OneThreadFileDividerTest.cs
@@ -8,15 +8,87 @@ namespace DHGSystems.FileSortingWithLimitedMemory
     public class OneThreadFileDividerTest
     {
         private readonly string _tempPath = "OneThreadFileDividerTestFolder";
+        private readonly string testFolderPath = @"TestFiles\\";
+        private readonly string oneRowTestFile = @"TestFiles\\oneLineFile.txt";
+        private readonly string emailTestFile = @"TestFiles\\EmailTest.txt";
+        private readonly string oneRowTestResultFile = @"TestFilesSorted\\oneLineFileSorted.txt";
+        private readonly string emailTestResultFile = @"TestFilesSorted\\emailTestSorted.txt";
 
         [TestMethod]
         public void ProcesOneRowFile_Should_BePositive()
         {
-            Directory.Delete(_tempPath, true);
-            Directory.CreateDirectory(_tempPath);
-            OneThreadFileDivider oneThreadFileDivider = new OneThreadFileDivider(_tempPath, 1, "sorted_file_", new DhgSystemsNLogLogger());
-            var result = oneThreadFileDivider.DivideFileWithSort("OneRowFile.txt", 1);
+            if (Directory.Exists(_tempPath))
+            {
+                Directory.Delete(_tempPath, true);
+            }
 
+            Directory.CreateDirectory(_tempPath);
+            OneThreadFileDivider oneThreadFileDivider =
+                new OneThreadFileDivider(_tempPath,  "sorted_file_", new DhgSystemsNLogLogger());
+            var generatedFiles = oneThreadFileDivider.DivideFileWithSort(oneRowTestFile, 5).ToList();
+            var fileContent = File.ReadAllText(generatedFiles.First());
+            var resultFileContent = File.ReadAllText(oneRowTestResultFile);
+            Assert.AreEqual(resultFileContent,fileContent);
+        }
+
+        [TestMethod]
+        public void ProcesMultiLineRowFile_Should_BePositive()
+        {
+            if (Directory.Exists(_tempPath))
+            {
+                Directory.Delete(_tempPath, true);
+            }
+
+            Directory.CreateDirectory(_tempPath);
+            OneThreadFileDivider oneThreadFileDivider =
+                new OneThreadFileDivider(_tempPath,  "sorted_file_", new DhgSystemsNLogLogger());
+            var generatedFiles = oneThreadFileDivider.DivideFileWithSort(emailTestFile, 40).ToList();
+            var fileContent = File.ReadAllText(generatedFiles.First());
+            var resultFileContent = File.ReadAllText(emailTestResultFile);
+            Assert.AreEqual(resultFileContent, fileContent);
+        }
+
+        [TestMethod]
+        public void ProcesMultiLineRowFile_OnLineCount_Equal_MaxBatch_Should_BePositive()
+        {
+            if (Directory.Exists(_tempPath))
+            {
+                Directory.Delete(_tempPath, true);
+            }
+
+            Directory.CreateDirectory(_tempPath);
+            OneThreadFileDivider oneThreadFileDivider =
+                new OneThreadFileDivider(_tempPath, "sorted_file_", new DhgSystemsNLogLogger());
+            var generatedFiles = oneThreadFileDivider.DivideFileWithSort(emailTestFile, 40).ToList();
+            var fileContent = File.ReadAllText(generatedFiles.First());
+            var resultFileContent = File.ReadAllText(emailTestResultFile);
+            Assert.AreEqual(resultFileContent, fileContent);
+        }
+
+
+
+
+        [DataTestMethod]
+        [DataRow(1, 5)]
+        [DataRow(2, 3)]
+        [DataRow(3, 2)]
+        [DataRow(4, 2)]
+        [DataRow(5, 1)]
+        [DataRow(6, 1)]
+
+        public void ProcesMultiLineFile_WithDifferent_DivideLineCount_Should_BePositive(long maxLineCount,
+            int expectedFileCount)
+        {
+            if (Directory.Exists(_tempPath))
+            {
+                Directory.Delete(_tempPath, true);
+            }
+
+            Directory.CreateDirectory(_tempPath);
+            OneThreadFileDivider oneThreadFileDivider =
+                new OneThreadFileDivider(_tempPath, "sorted_file_", new DhgSystemsNLogLogger());
+            var generatedFiles = oneThreadFileDivider.DivideFileWithSort(emailTestFile, maxLineCount).ToList();
+            Assert.AreEqual(expectedFileCount, generatedFiles.Count);
         }
     }
 }
